requirements.txt
flask==2.1.1
psycopg2==2.9.3
pyjwt==2.3.0